Rename 'medea_callback' as 'callback' and 'medea' as 'api' [run ci]
