+++ mock/control-api/src/api/mod.rs
@@ -16,7 +16,7 @@ use actix_web::{
     App, HttpResponse, HttpServer,
 };
 use clap::ArgMatches;
-use medea_control_api_proto::grpc::medea as proto;
+use medea_control_api_proto::grpc::api as proto;
 use serde::{Deserialize, Serialize};
 
 use crate::{
