+++ src/api/control/endpoints/mod.rs
@@ -8,7 +8,7 @@ pub mod webrtc_publish_endpoint;
 use std::convert::TryFrom;
 
 use derive_more::{Display, From, Into};
-use medea_control_api_proto::grpc::medea as proto;
+use medea_control_api_proto::grpc::api as proto;
 use serde::Deserialize;
 
 use super::{member::MemberElement, TryFromProtobufError};
