+++ src/api/control/endpoints/webrtc_play_endpoint.rs
@@ -5,7 +5,7 @@
 use std::convert::TryFrom;
 
 use derive_more::{Display, From, Into};
-use medea_control_api_proto::grpc::medea as proto;
+use medea_control_api_proto::grpc::api as proto;
 use serde::Deserialize;
 
 use crate::api::control::{refs::SrcUri, TryFromProtobufError};
