+++ src/api/control/error_codes.rs
@@ -7,7 +7,7 @@
 use std::string::ToString;
 
 use derive_more::Display;
-use medea_control_api_proto::grpc::medea as proto;
+use medea_control_api_proto::grpc::api as proto;
 
 use crate::{
     api::control::{
